Split middleware so decryption happen after axios
