+++ src/frontend/web_application/src/store/middlewares/encryption-middleware.js
@@ -1,14 +1,11 @@
-// Renaming REQUEST_DRAFT_SUCCESS actions for disambiguation.
-import { decryptDraftSuccess, REQUEST_DRAFT_SUCCESS as DRAFT_REQUEST_DRAFT_SUCCESS } from '../modules/draft-message';
-import { CREATE_MESSAGE, UPDATE_MESSAGE, FETCH_MESSAGES_SUCCESS, REQUEST_MESSAGE_SUCCESS, REQUEST_MESSAGES_SUCCESS, REQUEST_DRAFT_SUCCESS as MESSAGE_REQUEST_DRAFT_SUCCESS } from '../modules/message';
-// eslint-disable-next-line no-unused-vars
-import { askPassphrase, needPassphrase, needPrivateKey, encryptMessage as encryptMessageStart, encryptMessageSuccess, encryptMessageFail, decryptMessage as decryptMessageStart, decryptMessageSuccess, decryptMessageFail, SET_PASSPHRASE } from '../modules/encryption';
+import { CREATE_MESSAGE, UPDATE_MESSAGE } from '../modules/message';
+import { encryptMessage as encryptMessageStart, encryptMessageSuccess } from '../modules/encryption';
 import { requestRemoteIdentity } from '../modules/remote-identity';
 import { tryCatchAxiosAction } from '../../services/api-client';
-import { getKeysForEmail, getKeysForMessage, PUBLIC_KEY } from '../../services/openpgp-keychain-repository';
+import { getKeysForEmail, PUBLIC_KEY } from '../../services/openpgp-keychain-repository';
 import { getParticipantsKeys } from '../../modules/encryption/services/keyring/remoteKeys';
 import { identitiesSelector } from '../selectors/identities';
-import { isMessageEncrypted, decryptMessage, encryptMessage } from '../../services/encryption';
+import { encryptMessage } from '../../services/encryption';
 import { getAuthor } from '../../services/message';
 
 const getIdentities = (state, identitiesIds) =>
@@ -105,119 +102,11 @@ const encryptMessageAction = async (store, dispatch, action) => {
   return action;
 };
 
-const extractMessagesFromAction = ({ payload, type }) => {
-  switch (type) {
-    case MESSAGE_REQUEST_DRAFT_SUCCESS:
-    case DRAFT_REQUEST_DRAFT_SUCCESS:
-      return [payload.draft];
-    case REQUEST_MESSAGE_SUCCESS:
-      return [payload.data];
-    case REQUEST_MESSAGES_SUCCESS:
-    case FETCH_MESSAGES_SUCCESS:
-      return payload.data.messages;
-    default:
-      return [];
-  }
-};
-
-const getKeyPassphrase = (state, fingerprint) => {
-  const { privateKeysByFingerprint } = state.encryption;
-
-  return privateKeysByFingerprint[fingerprint] &&
-    privateKeysByFingerprint[fingerprint].status === 'ok' &&
-    privateKeysByFingerprint[fingerprint].passphrase;
-};
-
-const decryptMessageAction = async (state, dispatch, message) => {
-  if (!isMessageEncrypted(message)) {
-    return message;
-  }
-
-  // try {
-  const keys = await getKeysForMessage(message);
-
-  if (keys.length <= 0) {
-    dispatch(needPrivateKey({ message }));
-
-    return message;
-  }
-
-  let usableKey = keys.find(key => key.isDecrypted());
-  let passphrase = null;
-
-  if (!usableKey) {
-    usableKey = keys.find(key => getKeyPassphrase(state, key.getFingerprint()));
-
-    if (usableKey) {
-      passphrase = getKeyPassphrase(state, usableKey.getFingerprint());
-      await usableKey.decrypt(passphrase);
-    }
-  }
-
-  if (!usableKey) {
-    keys.forEach(key => dispatch(askPassphrase({ fingerprint: key.getFingerprint() })));
-    dispatch(needPassphrase({ message, fingerprints: keys.map(key => key.getFingerprint()) }));
-
-    return message;
-  }
-
-  dispatch(decryptMessageStart({ message }));
-  const decryptedMessage = await decryptMessage(message, [usableKey]);
-  dispatch(decryptMessageSuccess({ message, decryptedMessage }));
-
-  return decryptedMessage;
-  // } catch (e) {
-  // const { message: error } = e;
-  // dispatch(decryptMessageFail({ message, error }));
-
-  // return message;
-  // }
-};
-
-const decryptMessagesAction = async (state, dispatch, messages) => {
-  if (messages.length <= 0) return messages;
-
-  return Promise.all(messages.map(message => decryptMessageAction(state, dispatch, message)));
-};
-
-const findMessagesEncryptedWithKey = (state, fingerprint) => {
-  const { messageEncryptionStatusById } = state.encryption;
-
-  return Object.values(messageEncryptionStatusById)
-    .filter(messageEntry => fingerprint === messageEntry.keyFingerprint);
-};
-
-const setPassphraseAction = (state, dispatch, action) => {
-  const messages = findMessagesEncryptedWithKey(state, action.payload.fingerprint);
-
-  decryptMessagesAction(state, dispatch, messages);
-};
-
 export default store => next => async (action) => {
   switch (action.type) {
     case CREATE_MESSAGE:
     case UPDATE_MESSAGE:
       return next(await encryptMessageAction(store, store.dispatch, action));
-    case MESSAGE_REQUEST_DRAFT_SUCCESS:
-    case DRAFT_REQUEST_DRAFT_SUCCESS:
-      decryptMessageAction(store.getState(), store.dispatch, extractMessagesFromAction(action)[0])
-        .then(draft =>
-          store.dispatch(decryptDraftSuccess({ internalId: action.payload.internalId, draft })));
-
-      return next(action);
-    case REQUEST_MESSAGE_SUCCESS:
-    case REQUEST_MESSAGES_SUCCESS:
-    case FETCH_MESSAGES_SUCCESS:
-      decryptMessagesAction(store.getState(), store.dispatch, extractMessagesFromAction(action));
-
-      return next(action);
-    case SET_PASSPHRASE:
-      store.dispatch(action);
-      setPassphraseAction(store.getState(), store.dispatch, action);
-      // decrypt message
-      // init setTimeout to reset key
-      // forward passphrase to redux.
-      break;
     default:
       break;
   }
